No latest tag until a latest tag is published
With the current setup, when you are first pushing releases with a --tag=rc for example, sinopia insists that there is a latest tag.
I believe checking for the igore_latest_tag is enough to get the old sinopia behavior back. No need to add the latest dist-tag if non latest version was added.